+++ src/app/modules/data-query/data-query/cohort-definition/action-bar/cohort-definition-action-bar.component.ts
@@ -1,7 +1,7 @@
 import { Component, OnInit } from '@angular/core';
 import { FeasibilityQueryFactoryService } from 'src/app/service/FeasibilityQueryFactory.service';
 import { FeasibilityQueryProviderService } from 'src/app/service/Provider/FeasibilityQueryProvider.service';
-import { FeasibilityQueryValidationService } from 'src/app/service/Criterion/FeasibilityQueryValidation.service';
+import { FeasibilityQueryValidationService } from 'src/app/service/FeasibilityQuery/FeasibilityQueryValidation.service';
 import { NavigationHelperService } from 'src/app/service/NavigationHelper.service';
 import { Observable, Subscription } from 'rxjs';
 import { ResultProviderService } from 'src/app/service/Provider/ResultProvider.service';
