+++ src/app/modules/feasibility-query/components/editor/reference-criteria-modal/edit-reference-criteria-modal.component.ts
@@ -1,15 +1,15 @@
 import { AttributeFilter } from 'src/app/model/FeasibilityQuery/Criterion/AttributeFilter/AttributeFilter';
+import { BuildReferenceCriterionService } from 'src/app/service/Criterion/Build/BuildReferenceCriterionService';
 import { Component, HostListener, Inject, OnInit } from '@angular/core';
-import { CreateCriterionService } from 'src/app/service/Criterion/CreateCriterionService';
 import { Criterion } from 'src/app/model/FeasibilityQuery/Criterion/Criterion';
 import { CriterionBuilder } from 'src/app/model/FeasibilityQuery/Criterion/CriterionBuilder';
-import { CriterionValidationService } from '../../../../../service/Criterion/CriterionValidation.deprecated.service';
+import { CriterionValidationService } from 'src/app/service/Criterion/Validation/CriterionValidation.deprecated.service';
 import { Display } from 'src/app/model/DataSelection/Profile/Display';
+import { FilterTypes } from '../../../../../model/Utilities/FilterTypes';
 import { MAT_DIALOG_DATA, MatDialogRef } from '@angular/material/dialog';
 import { ReferenceCriterion } from 'src/app/model/FeasibilityQuery/Criterion/ReferenceCriterion';
 import { ReferenceCriterionProviderService } from 'src/app/service/Provider/ReferenceCriterionProvider.service';
 import { TerminologyCode } from 'src/app/model/Terminology/TerminologyCode';
-import { FilterTypes } from '../../../../../model/Utilities/FilterTypes';
 @Component({
   selector: 'num-edit-reference-criteria',
   templateUrl: './edit-reference-criteria-modal.component.html',
@@ -26,7 +26,7 @@ export class EditReferenceCriteriaModalComponent implements OnInit {
   constructor(
     @Inject(MAT_DIALOG_DATA) public data: EditReferenceCriteriaModalComponent,
     private dialogRef: MatDialogRef<EditReferenceCriteriaModalComponent, Criterion>,
-    private createCriterionService: CreateCriterionService,
+    private buildReferenceCriterionService: BuildReferenceCriterionService,
     private referenceCriterionProvider: ReferenceCriterionProviderService,
     private criterionValidationService: CriterionValidationService
   ) {}
@@ -46,8 +46,8 @@ export class EditReferenceCriteriaModalComponent implements OnInit {
   }
 
   public saveReferenceCriterion() {
-    this.createCriterionService
-      .createReferenceCriteriaFromHashes(this.ids, this.criterion.getId())
+    this.buildReferenceCriterionService
+      .buildReferenceCriteriaFromHashes(this.ids, this.criterion.getId())
       .subscribe((referenceCriteria: ReferenceCriterion[]) => {
         referenceCriteria.forEach((referenceCriterion) =>
           this.referenceCriterionProvider.setReferenceCriterionByUID(
