+++ src/app/service/DataQuery/DataQueryValidation.service.ts
@@ -1,7 +1,7 @@
 import { combineLatest, map, Observable } from 'rxjs';
 import { DataSelection } from 'src/app/model/DataSelection/DataSelection';
 import { DataSelectionProviderService } from 'src/app/modules/data-selection/services/DataSelectionProvider.service';
-import { FeasibilityQueryValidationService } from '../Criterion/FeasibilityQueryValidation.service';
+import { FeasibilityQueryValidationService } from '../FeasibilityQuery/FeasibilityQueryValidation.service';
 import { Injectable } from '@angular/core';
 import { ValidDataQuery } from 'src/app/model/Types/ValidDataQuery';
